[experimental] Add posibility to influence scoring of results
Basic, first-step implementation of a way to influence result scoring
based on user-specific values.

In this implemention users are able to mark results that are important
for them. These results will then get boosted in further searches by
them. There is also implementation for negative boosting - mark results
you want to supress - but it is disabled for now.

In this implementaion all results user marked are independant from
other users, and will not influence other user's results.

Change-Id: Id491499b3c71d270ebc198662d4932d3171499d1
ERM: #9775