Switch to single backend
There is no point in having additional complexity for multi-backend
structure, since we will probably never use more than a single backend.

This allows for much simpler overriding of host and port for the
ElasicSearch.
Also, Source config is moved to attribute, and, with the help of new
SourceRegistry, its much easier to add/extends sources, without the need
for messy hook

Change-Id: I55115b21d806629949d758b37c2e8b2d39e5740b