Remove server-side AC scoring
We are not using scorless suggest query for AC, so no need to do the
scoring server-side

Refine top-match logic. Top match will only be shown if the page is
exact match AND its hightest ranking. This is to avoid placing exact
matches from non-preferred NSs to the top.

Also, index page language

Change-Id: I8bb61b0bbd2cc80e8b1d1869593887a4530d3535
ERM: #10161