Page language and display title
- Index display title. Its now searchable and is shown in the page
titles along with original page title

- Index and filter by page language. Users can now filter pages in
particular language only. Also, if its so configured, language filter
can be set automatically based on user's language. This filter is
special, because when language filter is set, results will show all page
is this language + all the results that language is not applicable to
(files, SPs...).

Change-Id: I41b06c0d9d04e2d598f32952f1b18a0e6a038b21
ERM: #10981