Fallback for document indexing limits
With some documents it happens that tika limit of 100.000 characters is reached.
However, it appears to not be really related to char count, since plain text PDFs
with more than 2 mil. chars index fine. Since it cannot be reliably reproduced
and happens only sometimes, best we can do is to have a fallback.

This will try normal indexing, if it fails it will index only file, without content,
so that its, at least, in the index and be found and located.

[3.1.1]

Change-Id: I163c1f377c328c46cafee35f75aaa1990cfcf92b
ERM: #15725